backend-drf/requirements.txt
absl-py==2.3.1
asgiref==3.9.2
astunparse==1.6.3
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.15.0
idna==3.11
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
numpy==2.3.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
peewee==3.18.2
pillow==11.3.0
platformdirs==4.5.0
protobuf==6.32.1
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yfinance==0.2.66
